pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecommerce-price-monitor"
version = "1.0.0"
description = "A comprehensive e-commerce price monitoring and analysis system"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ecommerce", "price", "monitoring", "analysis", "scraping"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "selenium>=4.8.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.11.0",
    "plotly>=5.11.0",
    "openpyxl>=3.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=0.19.0",
    "schedule>=1.1.0",
    "rich>=12.0.0",
    "click>=8.1.0",
    "lxml>=4.9.0",
    "aiohttp>=3.8.0",
    "asyncio>=3.4.3",
    "jieba>=0.42.1",
    "zhconv>=1.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
]

[project.scripts]
ecommerce-monitor = "ecommerce_price_monitor.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/ecommerce-price-monitor"
Documentation = "https://github.com/yourusername/ecommerce-price-monitor#readme"
Repository = "https://github.com/yourusername/ecommerce-price-monitor"
"Bug Tracker" = "https://github.com/yourusername/ecommerce-price-monitor/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=ecommerce_price_monitor --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
source = ["src/ecommerce_price_monitor"]
omit = ["*/tests/*", "*/test_*.py"]